feat: item v5
