+++ lib/atol/version.rb
@@ -2,6 +2,12 @@
 
 module Atol
   module Version
+    V4 = 'https://online.atol.ru/possystem/v4'
+    V5 = 'https://online.atol.ru/possystem/v5'
+
+    V4_TEST = 'https://testonline.atol.ru/possystem/v4'
+    V5_TEST = 'https://testonline.atol.ru/possystem/v5'
+
     LIB = '1.1.0'
   end
 end
